Add files via upload
Below are the .html files to my web design and development cw. I was student 1 and was asked to create a splash screen along with the shop of the website our group chose. In addition I also added a contents page and a page editor for easier understanding of my work.